verify the rpm pacakge after install
verify package libvirt-daemon-config-nwfilter
