--- libvirt/tests/cfg/nwfilter/nwfilter_binding_create.cfg
@@ -1,13 +1,0 @@
-- nwfilter_binding_create:
-    type = nwfilter_binding_create
-    start_vm = "yes"
-    status_error = "no"
-    kill_vm = "yes"
-    variants:
-        - check:
-            check_cmd = "ebtables -t nat -L"
-            check_filter = "nwfilter-binding-list"
-            filter_name = "clean-traffic"
-            wait_time = "1"
-            expected_match = "-j DROP"
-            cmd_restart = "systemctl restart libvirtd"
