+++ libvirt/tests/cfg/nwfilter/verify_libvirt_daemon_config_nwfilter.cfg
@@ -1,0 +1,6 @@
+- verify_libvirt_daemon_config_nwfilter:
+    type = verify_libvirt_daemon_config_nwfilter
+    start_vm = "no"
+    kill_vm = "no"
+    status_error = "no"
+    check_cmd = "rpm -V libvirt-daemon-config-nwfilter"
